Split datagen into separate source set.
