+++ src/datagen/kotlin/me/jcotton42/alloyance/datagen/AlloyanceDataMapProvider.kt
