+++ src/datagen/kotlin/me/jcotton42/alloyance/datagen/AlloyanceRecipesProvider.kt
