+++ src/datagen/kotlin/me/jcotton42/alloyance/datagen/SimpleRecipeBuilder.kt
